code/server/requirements.txt
acres==0.3.0
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
dnspython==2.7.0
docx2txt==0.8
email_validator==2.2.0
etelemetry==0.3.1
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
frontend==0.0.3
fsspec==2025.3.0
google==3.0.0
google-auth==2.38.0
google-genai==1.7.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pillow==11.1.0
prov==2.0.1
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.4
pyparsing==3.2.2
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytils==0.4.1
pytz==2025.1
pyxnat==1.6.3
PyYAML==6.0.2
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
setuptools==77.0.3
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.1
tokenizers==0.21.1
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
transformers==4.50.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
wheel==0.45.1
